Inverse relations added both to the detail and edit page

+++ server/Conscience/Conscience.Application/Graph/Entities/ConscienceMutationBase.cs
@@ -12,16 +12,19 @@ namespace Conscience.Application.Graph.Entities
 {
     public abstract class ConscienceMutationBase : ObjectGraphType<object>
     {
-        protected void ModifyCollection<T, R>(BaseRepository<R> repository, LogEntryService logService, ICollection<T> collection, ICollection<T> dbCollection, Action<T, T> setProperties, Func<T, T, bool> additionalPredicate = null)
+        protected void ModifyCollection<T, R>(BaseRepository<R> repository, LogEntryService logService, ICollection<T> collection, ICollection<T> dbCollection, Action<T, T> setProperties, Func<T, T, bool> additionalPredicate = null, bool shouldRemove = true)
             where T : IdentityEntity, new()
             where R : class
         {
-            var itemsToRemove = dbCollection.Where(dbItem => !collection.Any(item => dbItem.Id == item.Id || (additionalPredicate != null && additionalPredicate(item, dbItem)))).ToList();
-            foreach (var itemToRemove in itemsToRemove)
+            if (shouldRemove)
             {
-                logService.Log(null, $"Removed {typeof(T).Name} '{itemToRemove.ToString()}'");
-                dbCollection.Remove(itemToRemove);
-                repository.DeleteChild(itemToRemove);
+                var itemsToRemove = dbCollection.Where(dbItem => !collection.Any(item => dbItem.Id == item.Id || (additionalPredicate != null && additionalPredicate(item, dbItem)))).ToList();
+                foreach (var itemToRemove in itemsToRemove)
+                {
+                    logService.Log(null, $"Removed {typeof(T).Name} '{itemToRemove.ToString()}'");
+                    dbCollection.Remove(itemToRemove);
+                    repository.DeleteChild(itemToRemove);
+                }
             }
 
             foreach (var item in collection)
